% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaseq_is_valid.R
\name{aaseq_is_valid}
\alias{aaseq_is_valid}
\title{Check protein sequences for non-standard amino acids}
\usage{
aaseq_is_valid(seq)
}
\arguments{
\item{seq}{A vector of protein sequences}
}
\value{
A logical vector where TRUE indicates a valid protein sequence and FALSE indicates a sequence with invalid amino acids
}
\description{
Any proteins that contains an amino acid that is not one of the 20 standard amino acids is flagged as invalid
}
\examples{

# non_standard_df <- readRDS(system.file("extdata/non_standard_df.rds", package = "ampir"))

## Example dataframe (non_standard_df)
# non_standard_df
#       seq_name            seq_aa
# [1] G1P6H5_MYOLU    MALTVRIQAACLLLLLLASLTSYSLLLSQTTQLADLQTQ....
# [2] fake_sequence   MKVTHEUSYR$GXMBIJIDG*M80-\%

# aaseq_is_valid(non_standard_df$seq.aa)

## Output
# [1]  TRUE FALSE
}
