% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampir.R
\docType{package}
\name{ampir}
\alias{ampir}
\alias{ampir-package}
\title{ampir: Fast and accurate prediction of antimicrobial peptides}
\description{
Some general text about ampir here
}
\section{Preparing data}{

Use \code{\link{read_faa}} to read from FASTA
Use \code{\link{calculate_features}} to calculate features
}

\section{Prediction}{

Use \code{\link{predict_AMP_prob}} to predict amps
}

