% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pseudo_comp.R
\name{calc_pseudo_comp}
\alias{calc_pseudo_comp}
\title{Calculate the pseudo amino acid composition}
\usage{
calc_pseudo_comp(seq, lambda = NULL)
}
\arguments{
\item{seq}{A protein sequence}

\item{lambda}{lambda value (determined within the function)}
}
\description{
This function is adapted from the extractPAAC function from the protr package (https://github.com/nanxstats/protr)
}
\references{
Nan Xiao, Dong-Sheng Cao, Min-Feng Zhu, and Qing-Song Xu. (2015). protr/ProtrWeb: R package and web server for generating various numerical representation schemes of protein sequences. Bioinformatics 31 (11), 1857-1859.
}
