% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Calculate a set of numerical features from protein sequences}
\usage{
calculate_features(seq)
}
\arguments{
\item{seq}{A single protein sequence or dataframe column containing multiple protein sequences}
}
\value{
A dataframe containing numerical values related to the protein features
}
\description{
This function calculates set physicochemical and compositional features from protein sequences
}
\note{
This function depends on the Peptides package
}
\examples{
# Calculate features from Hepcidin AMP from Myotis lucifugus (UniProt ID G1P6H5)
calculate_features(seq =
   "MALTVRIQAACLLLLLLASLTSYSLLLSQTTQLADLQTQDTAGATAGLMPGLQRRRRRDTHFPICIFCCGCCYPSKCGICCKT")

## Output (showing the first five output columns)
# Length  Amphiphilicity  Hydrophobicity     pI          Mw         ....
# [1] 83      0.4145847       0.4373494     8.501312     9013.757   ....
}
\references{
Osorio, D., Rondon-Villarreal, P. & Torres, R. Peptides: A package for data mining of antimicrobial peptides. The R Journal. 7(1), 4–14 (2015).
}
