% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_rows.R
\name{chunk_rows}
\alias{chunk_rows}
\title{Determine row breakpoints for dividing a dataset into chunks for parallel processing}
\usage{
chunk_rows(nrows, n_cores)
}
\arguments{
\item{nrows}{The number of rows in the dataset to be chunked}

\item{n_cores}{The number of cores that will be used for parallel processing}
}
\value{
A list of integer vectors consisting of the rows in each chunk
}
\description{
Determine row breakpoints for dividing a dataset into chunks for parallel processing
}
