% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_faa.R
\name{df_to_faa}
\alias{df_to_faa}
\title{Save a dataframe in FASTA format}
\usage{
df_to_faa(df, file = "")
}
\arguments{
\item{df}{a dataframe containing two columns (seq.name and seq.aa)}

\item{file}{file path to save the named file to}
}
\value{
A FASTA file where protein sequences are represented in two lines: The protein name preceded by a greater than symbol,
        and a new second line that contains the protein sequence
}
\description{
This function writes a dataframe out as a FASTA format file
}
\examples{

# Use \\code{read_faa} to read a FASTA file as a dataframe
my_protein <- read_faa(system.file("extdata/bat_protein.fasta", package = "ampir"))

# Use \\code{df_to_faa} to write a dataframe into FASTA file format
df_to_faa(my_protein,(system.file("extdata/my_protein.fasta", package = "ampir")))


## Output written in "my_protein.fasta"
#[1] >G1P6H5_MYOLU
#[2] MALTVRIQAACLLLLLLASLTSYSLLLSQTTQLADLQTQDTAGATAGLMPGLQRRRRRDTHFPICIFCCGCCYPSKCGICCKT
}
