% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_AMP_prob.R
\name{predict_AMP_prob}
\alias{predict_AMP_prob}
\title{Predict the antimicrobial peptide probability of a protein}
\usage{
predict_AMP_prob(df)
}
\arguments{
\item{df}{A dataframe containing numerical features (as calculated by \code{calculate_features})}
}
\value{
A dataframe containing a single column with probability values
}
\description{
This function predicts the probability of a protein to be an antimicrobial peptide based on feature calculations (as obtained from \code{calculate_features})
}
\note{
This model was created via the caret package (https://github.com/topepo/caret/)
}
\examples{

predict_AMP_prob(bat_features)
 # prob_AMP
[1] 0.9672241
}
