% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_aas.R
\name{random_aa}
\alias{random_aa}
\title{Create a random amino acid sequence of given length}
\usage{
random_aa(x)
}
\arguments{
\item{x}{Sequence length for random_aa}
}
\value{
A character vector of given length
}
\description{
These functions create random amino acid sequences.
random_aa results in a single sequence of any given length,
random_aas results in a given number of amino acid sequences with a default random length range between 10 and 2000.
}
\examples{
random_aa(x = 10)
[1] "IIANHDNADR"

}
