% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_aas.R
\name{random_aas}
\alias{random_aas}
\title{Create multiple random amino acid sequences of given length}
\usage{
random_aas(n, min_length = 10, max_length = 2000)
}
\arguments{
\item{n}{The number of sequences}

\item{min_length, max_length}{The minimum and maximum length of the sequence (default is 10 and 2000, respectively)}
}
\value{
A character vector of given number and length range
}
\description{
This function uses a helper function (see \code{random_aa}) and creates random amino acid sequences of a given number and length range
}
