% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonstandard_aa.R
\name{remove_nonstandard_aa}
\alias{remove_nonstandard_aa}
\title{Remove non standard amino acids from protein sequences}
\usage{
remove_nonstandard_aa(df, seq)
}
\arguments{
\item{df}{a dataframe}

\item{seq}{a column in a dataframe containing protein sequences}
}
\value{
a dataframe like the input dataframe but with removed proteins that contained non standard amino acids
}
\description{
This function removes anything that is not one of the 20 standard amino acids in protein sequences
}
\examples{
## Example dataframe (param df input)
# seq.name        seq.aa
# G1P6H5_MYOLU    MALTVRIQAACLLLLLLASLTSYSLLLSQTTQLADLQTQ....
# fake_sequence   MKVTHEUSYR$GXMBIJIDGM80-\%

remove_nonstandard_aa(df, df$seq.aa)

## Output
# seq.name        seq.aa
# G1P6H5_MYOLU    MALTVRIQAACLLLLLLASLTSYSLLLSQTTQLADLQTQ....
}
