% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stop_codon.R
\name{remove_stop_codon}
\alias{remove_stop_codon}
\title{Remove stop codon at end of sequence}
\usage{
remove_stop_codon(faa_df)
}
\arguments{
\item{faa_df}{A dataframe containing two columns: the sequence name and amino acid sequence}
}
\value{
The input dataframe without the stop codons at the end of sequences
}
\description{
Stop codons at the end of the amino acid sequences are removed
}
