% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pseudo_comp.R
\name{calc_pseudo_comp}
\alias{calc_pseudo_comp}
\title{Calculate the pseudo amino acid composition}
\usage{
calc_pseudo_comp(seq)
}
\arguments{
\item{seq}{A protein sequence}
}
\description{
Calculate the pseudo amino acid composition
}
\references{
Nan Xiao, Dong-Sheng Cao, Min-Feng Zhu, and Qing-Song Xu. (2015). protr/ProtrWeb: R package and web server for generating various numerical representation schemes of protein sequences. Bioinformatics 31 (11), 1857-1859.
The imported function within this function originates from the protr package (https://github.com/nanxstats/protr)
}
