% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_aas.R
\name{random_aas}
\alias{random_aas}
\title{Create multiple random amino acid sequences of given length}
\usage{
random_aas(n, min_length = 10, max_length = 2000)
}
\arguments{
\item{n}{The number of sequences for random_aas}

\item{min_length, max_length}{The minimum and maximum length of the sequence (default is 10 and 2000, respectively) for random_aas

random_aas(n = 3, min_length = 5, max_length = 20)
[1] "IGPRWSFRLW"   "NEAHNW"  "LDNCQRTLWKFPHHFHL"}
}
\description{
Create multiple random amino acid sequences of given length
}
